pyproject.toml
[project]
name = "hr-tooling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "passlib[bcrypt]",
    "PyJWT",
]
